#UI Priority #Store Sell #JumpBugFix
유아이 우선출력순위 설정
상점 판매기능 추가 / 버그수정
점프버그픽스

+++ MapleStory 2008/MapleStory/RenderMgr.h
@@ -5,7 +5,7 @@ class CRenderMgr
 {
 private:
 	static CRenderMgr*		m_pInstance;
-	multimap<float, CUI*>	m_MapRenderUI;
+	multimap<int, CUI*>	m_MapRenderUI;
 
 public:
 	void	AddUI(CUI* pUI);
